Add keydown handler to editable element and handle enter key for newlines
This approach turned out to be more difficult than I hoped because
of the odd behavior that must be emulated.

The desired behavior is exhibited when the code element has the
contenteditable attribute, and the user presses enter. The browser
will append two new text nodes to the code element, one with a
single newline character, and then another empty text node that
appears on a new line, with cursor focus.

When I attempted to replicate this behavior in a keydown handler,
I got very different behavior, even though I seemed to have replicated
the content of the child text nodes.

Further experimentation would be required to continue making progress.
But I think we will have better luck just moving the contenteditable
attribute to the code element.